feat(settings): split CSV import into its own 'Import Custom Food Data' menu entry
Two changes:

1. Renamed the existing 'Export / Import data' menu entry to
   'Export / Import App Data' to make it clear it operates on the
   app-wide ZIP archive (user activity, intakes, tracked days).

2. Extracted the Sample CSV download + Import meals (CSV) buttons into
   their own dialog (ImportCustomFoodDataDialog) reachable via a new
   'Import Custom Food Data' menu entry placed directly above the App
   Data entry. Keeps food-data import visually separate from the
   whole-app backup flow.

Both dialogs reuse ExportImportBloc — no new use cases, events, or
states required.

l10n: renamed exportImportLabel -> exportImportAppDataLabel and added
importCustomFoodDataLabel + importCustomFoodDataDescription across all
8 locales (en, de, cz, it, pl, tr, uk, zh).
