+++ lib/features/settings/presentation/widgets/export_import_dialog.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:opennutritracker/core/utils/locator.dart';
@@ -7,14 +6,8 @@ import 'package:opennutritracker/features/diary/presentation/bloc/diary_bloc.dar
 import 'package:opennutritracker/features/home/presentation/bloc/home_bloc.dart';
 import 'package:opennutritracker/features/settings/presentation/bloc/export_import_bloc.dart';
 import 'package:opennutritracker/generated/l10n.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 class ExportImportDialog extends StatelessWidget {
-  static final _offAndroidUrl = Uri.parse(
-      'https://play.google.com/store/apps/details?id=org.openfoodfacts.scanner');
-  static final _offIosUrl = Uri.parse(
-      'https://apps.apple.com/us/app/open-food-facts-product-scan/id588797948');
-
   final exportImportBloc = locator<ExportImportBloc>();
 
   final _homeBloc = locator<HomeBloc>();
@@ -27,14 +20,13 @@ class ExportImportDialog extends StatelessWidget {
   Widget build(BuildContext context) {
     return AlertDialog(
       title: Text(
-        S.of(context).exportImportLabel,
+        S.of(context).exportImportAppDataLabel,
         overflow: TextOverflow.ellipsis,
         maxLines: 2,
       ),
       content: Wrap(
         children: [
           Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               BlocBuilder<ExportImportBloc, ExportImportState>(
                 bloc: exportImportBloc,
@@ -70,22 +62,6 @@ class ExportImportDialog extends StatelessWidget {
                         Text(S.of(context).exportImportErrorLabel),
                       ],
                     );
-                  } else if (state is CsvImportResultState) {
-                    refreshScreens();
-                    return _buildCsvResult(context, state);
-                  } else if (state is CsvImportErrorState) {
-                    return Row(
-                      children: [
-                        Icon(
-                          Icons.error,
-                          color: Theme.of(context).colorScheme.error,
-                        ),
-                        const SizedBox(width: 8),
-                        Expanded(
-                          child: Text(S.of(context).csvImportErrorLabel),
-                        ),
-                      ],
-                    );
                   }
                   return const SizedBox.shrink();
                 },
@@ -103,72 +79,6 @@ class ExportImportDialog extends StatelessWidget {
           onPressed: () => exportImportBloc.add(ImportDataEvent()),
           child: Text(S.of(context).importAction),
         ),
-        TextButton(
-          onPressed: () => exportImportBloc.add(DownloadSampleCsvEvent()),
-          child: Text(S.of(context).downloadSampleCsvAction),
-        ),
-        TextButton(
-          onPressed: () => exportImportBloc.add(ImportMealsCsvEvent()),
-          child: Text(S.of(context).importMealsCsvAction),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildCsvResult(BuildContext context, CsvImportResultState state) {
-    final summary = state.skipped == 0
-        ? S.of(context).csvImportSuccessLabel(state.imported)
-        : S
-            .of(context)
-            .csvImportPartialLabel(state.imported, state.skipped);
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            Icon(
-              Icons.check_circle,
-              color: Theme.of(context).colorScheme.primary,
-            ),
-            const SizedBox(width: 8),
-            Expanded(child: Text(summary)),
-          ],
-        ),
-        if (state.anyHadBarcode) ...[
-          const SizedBox(height: 12),
-          Text.rich(
-            TextSpan(
-              children: [
-                TextSpan(
-                  text: S.of(context).csvImportContributeOffPrefix,
-                  style: Theme.of(context).textTheme.bodySmall,
-                ),
-                const TextSpan(text: ' '),
-                TextSpan(
-                  text: S.of(context).csvImportContributeOffAndroidLink,
-                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                        color: Theme.of(context).colorScheme.primary,
-                        decoration: TextDecoration.underline,
-                      ),
-                  recognizer: TapGestureRecognizer()
-                    ..onTap = () => launchUrl(_offAndroidUrl,
-                        mode: LaunchMode.externalApplication),
-                ),
-                const TextSpan(text: ' / '),
-                TextSpan(
-                  text: S.of(context).csvImportContributeOffIosLink,
-                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                        color: Theme.of(context).colorScheme.primary,
-                        decoration: TextDecoration.underline,
-                      ),
-                  recognizer: TapGestureRecognizer()
-                    ..onTap = () => launchUrl(_offIosUrl,
-                        mode: LaunchMode.externalApplication),
-                ),
-              ],
-            ),
-          ),
-        ],
       ],
     );
   }
