+++ lib/l10n/intl_cz.arb
@@ -83,7 +83,9 @@
   "customMealsEmptyLabel": "Zatím žádná vlastní jídla uložena.",
   "customMealsDeleteConfirmTitle": "Smazat vlastní jídlo?",
   "customMealsDeleteConfirmContent": "Všechny záznamy v deníku používající toto jídlo budou také odstraněny.",
-  "exportImportLabel": "Export / Import dat",
+  "exportImportAppDataLabel": "Export / Import dat aplikace",
+  "importCustomFoodDataLabel": "Importovat vlastní data potravin",
+  "importCustomFoodDataDescription": "Importujte vlastní jídla ze souboru CSV. Stáhněte si vzor, abyste viděli očekávaný formát sloupců a povinná pole.",
   "exportImportDescription": "Můžete uložit data z aplikace do .zip archívu a později je znovu importovat. To je užitečné, pokud potřebujete data zálohovat, nebo přenést na jiné zařízení.\n\nAplikace pro ukládání dat nepoužívá žádné cloudové služby.",
   "exportImportSuccessLabel": "Export / Import byl úspěšný",
   "exportImportErrorLabel": "Export / Import selhal",
