+++ lib/l10n/intl_de.arb
@@ -94,7 +94,9 @@
   "customMealsEmptyLabel": "Noch keine eigenen Mahlzeiten gespeichert.",
   "customMealsDeleteConfirmTitle": "Eigene Mahlzeit löschen?",
   "customMealsDeleteConfirmContent": "Alle Tagebucheinträge, die diese Mahlzeit verwenden, werden ebenfalls entfernt.",
-  "exportImportLabel": "Daten Exportieren / Importieren",
+  "exportImportAppDataLabel": "App-Daten exportieren / importieren",
+  "importCustomFoodDataLabel": "Eigene Lebensmittel-Daten importieren",
+  "importCustomFoodDataDescription": "Eigene Mahlzeiten aus einer CSV-Datei importieren. Beispiel herunterladen, um das erwartete Format und die Pflichtfelder zu sehen.",
   "exportImportDescription": "Sie können die App-Daten in eine Zip-Datei exportieren und später importieren. Dies ist nützlich, wenn Sie Ihre Daten sichern oder auf ein anderes Gerät übertragen möchten.\n\nDie App nutzt keinen Cloud-Dienst, um Ihre Daten zu speichern.",
   "exportImportSuccessLabel": "Export / Import erfolgreich",
   "exportImportErrorLabel": "Fehler beim Export/Import",
