+++ lib/l10n/intl_en.arb
@@ -96,7 +96,9 @@
   "customMealsEmptyLabel": "No custom meals saved yet.",
   "customMealsDeleteConfirmTitle": "Delete custom meal?",
   "customMealsDeleteConfirmContent": "All diary entries using this meal will also be removed.",
-  "exportImportLabel": "Export / Import data",
+  "exportImportAppDataLabel": "Export / Import App Data",
+  "importCustomFoodDataLabel": "Import Custom Food Data",
+  "importCustomFoodDataDescription": "Import your own meals from a CSV file. Download a sample to see the expected column shape and required fields.",
   "exportImportDescription": "You can export the app data to a zip file and import it later. This is useful if you want to backup your data or transfer it to another device.\n\nThe app does not use any cloud service to store your data.",
   "exportImportSuccessLabel": "Export / Import successful",
   "exportImportErrorLabel": "Export / Import error",
