+++ lib/l10n/intl_it.arb
@@ -83,7 +83,9 @@
   "customMealsEmptyLabel": "Nessun pasto personalizzato salvato.",
   "customMealsDeleteConfirmTitle": "Eliminare il pasto personalizzato?",
   "customMealsDeleteConfirmContent": "Tutte le voci del diario che utilizzano questo pasto verranno rimosse.",
-  "exportImportLabel": "Esporta / Importa dati",
+  "exportImportAppDataLabel": "Esporta / Importa dati app",
+  "importCustomFoodDataLabel": "Importa dati alimentari personalizzati",
+  "importCustomFoodDataDescription": "Importa i tuoi pasti da un file CSV. Scarica un esempio per vedere il formato delle colonne e i campi obbligatori.",
   "exportImportDescription": "Puoi esportare i dati dell'app in un file zip e importarli successivamente. Utile per backup o trasferimento su un altro dispositivo.\n\nL'app non utilizza servizi cloud per memorizzare i tuoi dati.",
   "exportImportSuccessLabel": "Esportazione / Importazione riuscita",
   "exportImportErrorLabel": "Errore di esportazione / importazione",
