+++ lib/l10n/intl_pl.arb
@@ -83,7 +83,9 @@
   "customMealsEmptyLabel": "Brak zapisanych własnych posiłków.",
   "customMealsDeleteConfirmTitle": "Usunąć własny posiłek?",
   "customMealsDeleteConfirmContent": "Wszystkie wpisy w dzienniku używające tego posiłku zostaną również usunięte.",
-  "exportImportLabel": "Eksport / Import danych",
+  "exportImportAppDataLabel": "Eksportuj / Importuj dane aplikacji",
+  "importCustomFoodDataLabel": "Importuj własne dane żywności",
+  "importCustomFoodDataDescription": "Importuj własne posiłki z pliku CSV. Pobierz przykład, aby zobaczyć oczekiwany format kolumn i wymagane pola.",
   "exportImportDescription": "Możesz wyeksportować dane aplikacji do pliku zip i zaimportować je później. Jest to przydatne, jeśli chcesz wykonać kopię zapasową danych lub przenieść je na inne urządzenie.\n\nAplikacja nie korzysta z żadnej usługi chmurowej do przechowywania danych.",
   "exportImportSuccessLabel": "Eksport / Import zakończony sukcesem",
   "exportImportErrorLabel": "Błąd eksportu / importu",
