+++ lib/l10n/intl_tr.arb
@@ -96,7 +96,9 @@
   "customMealsEmptyLabel": "Henüz özel yemek kaydedilmedi.",
   "customMealsDeleteConfirmTitle": "Özel yemek silinsin mi?",
   "customMealsDeleteConfirmContent": "Bu yemeği kullanan tüm günlük girişleri de kaldırılacak.",
-  "exportImportLabel": "Verileri Dışa Aktar / İçe Aktar",
+  "exportImportAppDataLabel": "Uygulama Verilerini Dışa / İçe Aktar",
+  "importCustomFoodDataLabel": "Özel Gıda Verilerini İçe Aktar",
+  "importCustomFoodDataDescription": "Kendi öğünlerinizi bir CSV dosyasından içe aktarın. Beklenen sütun yapısını ve zorunlu alanları görmek için örneği indirin.",
   "exportImportDescription": "Uygulama verilerini bir zip dosyasına dışa aktarabilir ve daha sonra içe aktarabilirsiniz. Bu, verilerinizi yedeklemek veya başka bir cihaza aktarmak istiyorsanız kullanışlıdır.\n\nUygulama, verilerinizi saklamak için herhangi bir bulut hizmeti kullanmaz.",
   "exportImportSuccessLabel": "Dışa Aktarma / İçe Aktarma başarılı",
   "exportImportErrorLabel": "Dışa Aktarma / İçe Aktarma hatası",
