+++ lib/l10n/intl_uk.arb
@@ -83,7 +83,9 @@
   "customMealsEmptyLabel": "Ще немає збережених власних страв.",
   "customMealsDeleteConfirmTitle": "Видалити власну страву?",
   "customMealsDeleteConfirmContent": "Усі записи в щоденнику, що використовують цю страву, також буде видалено.",
-  "exportImportLabel": "Експорт / Імпорт даних",
+  "exportImportAppDataLabel": "Експортувати / Імпортувати дані застосунку",
+  "importCustomFoodDataLabel": "Імпортувати власні дані про їжу",
+  "importCustomFoodDataDescription": "Імпортуйте власні страви з CSV файлу. Завантажте зразок, щоб побачити очікуваний формат стовпців та обов'язкові поля.",
   "exportImportDescription": "Ви можете експортувати дані додатка у zip-файл і імпортувати їх пізніше. Це корисно, якщо ви хочете зробити резервну копію або перенести дані на інший пристрій. Додаток не використовує жодних хмарних сервісів для зберігання ваших даних.",
   "exportImportSuccessLabel": "Експорт / Імпорт успішний",
   "exportImportErrorLabel": "Помилка експорту / імпорту",
