+++ lib/l10n/intl_zh.arb
@@ -84,7 +84,9 @@
   "customMealsEmptyLabel": "尚未保存自定义餐食。",
   "customMealsDeleteConfirmTitle": "删除自定义餐食？",
   "customMealsDeleteConfirmContent": "所有使用此餐食的日记条目也将被删除。",
-  "exportImportLabel": "导出/导入数据",
+  "exportImportAppDataLabel": "导出 / 导入应用数据",
+  "importCustomFoodDataLabel": "导入自定义食品数据",
+  "importCustomFoodDataDescription": "从 CSV 文件导入您自己的餐食。下载示例查看所需的列格式和必填字段。",
   "exportImportDescription": "您可以将应用数据导出到zip文件并稍后导入。这在您想要备份数据或传输到另一台设备时很有用。\n\n应用不会使用任何云服务存储您的数据。",
   "exportImportSuccessLabel": "导出/导入成功",
   "exportImportErrorLabel": "导出/导入错误",
